+++ Controls/_filterPopup/Panel.ts
@@ -4,7 +4,7 @@ import DetailPanel from 'Controls/_filterPopup/DetailPanel';
  * @class Controls/_filterPopup/Panel
  * @extends Controls/_filterPopup/DetailPanel
  * @public
- * @author Золотова Э.Е.
+ * @author Михайлов С.Е.
  * 
  * @deprecated Контрол устарел и в скором времени будет удалён, используйте {@link Controls/filterPopup:DetailPanel}.
  */
